Added as new class and added unittest
