+++ phpunit.xml
@@ -23,4 +23,7 @@
             <directory suffix=".php">./src</directory>
         </whitelist>
     </filter>
+    <php>
+         <ini name="display_errors" value="true"/>
+    </php>
 </phpunit>
