+++ src/CssInlinerPlugin.php
@@ -14,6 +14,7 @@ class CssInlinerPlugin implements \Swift_Events_SendListener
         $message = $evt->getMessage();
         
         $converter = new CssToInlineStyles();
+        $external = new ExternalCssPlugin();
         $converter->setEncoding($message->getCharset());
         $converter->setUseInlineStylesBlock();
         $converter->setCleanup();
@@ -21,7 +22,7 @@ class CssInlinerPlugin implements \Swift_Events_SendListener
         if ($message->getContentType() === 'text/html' ||
             ($message->getContentType() === 'multipart/alternative' && $message->getBody())
         ) {
-            $message->setBody($this->includeExternalStylesheets($message->getBody()), 'text/html');
+            $message->setBody($external->includeExternalStylesheets($message->getBody()), 'text/html');
             $converter->setHTML($message->getBody());
             $message->setBody($converter->convert());
         }
@@ -32,47 +33,6 @@ class CssInlinerPlugin implements \Swift_Events_SendListener
                 $part->setBody($converter->convert());
             }
         }
-    }
-
-    private function includeExternalStylesheets($message)
-    {
-        preg_match_all('/<link[^>]+>/mi', $message, $matches);
-        $linkElements = $matches[0];
-        $count = count($linkElements);
-
-        if ($linkElements) {
-            foreach ($linkElements as $i => $linkElement) {
-                $cssContent = $this->getCssContent($linkElement);
-                $message = str_replace($linkElement, '
-                    <style type="text/css">
-                        ' . $cssContent . '
-                    </style>
-                ', $message);
-            }
-        }
-        return $message;
-    }
-
-    /**
-     * @param $linkElement
-     * @return string
-     */
-    private function getCssContent($linkElement)
-    {
-        // load data from html element
-        preg_match_all("/href=['\"]([^'\"]+)['\"]/", $linkElement, $matches);
-        $url = $matches[1][0];
-
-        $absUrl = $url;
-        
-        // this needs to be changed to search in all include paths
-        $cssContent = @file_get_contents($absUrl);
-
-        if (!$cssContent) {
-            trigger_error('Error loading stylesheet from ' . $absUrl, E_USER_NOTICE);
-        }
-
-        return $cssContent;
     }
