+++ tests/stubs/converted-html-external.stub
@@ -1,0 +1,1 @@
+<html><head><style type="text/css">body {background: green;}</style></head></html>
