+++ tests/stubs/externalStyle.css
@@ -1,0 +1,1 @@
+body {background: green;}
