+++ tests/stubs/original-html-external.stub
@@ -1,0 +1,1 @@
+<html><head><link rel="stylesheet" href="%styleUrl%"></head></html>
